// modules/domain/teams/team.entity.d.ts
export declare class Team {
    id: string;
    provider: string;
    provider_id: number;
    name: string;
    name_zh: string;
    short_name: string;
    country: string;
    country_code: string;
    type: string;
    logo_url: string;
    venue: string;
    coach: string;
    founded: number;
    meta: Record<string, any>;
    created_at: Date;
    updated_at: Date;
}
//# sourceMappingURL=team.entity.d.ts.map